requirements.txt
fastapi==0.115.11
google-cloud-documentai==3.2.0
google-generativeai==0.8.4
python-multipart==0.0.20
uvicorn==0.34.0
python-dotenv==1.0.1
pillow==11.1.0